IGS Informatika/Informatika_11 Mr. Anas/SQLite/MEET03 G-Sheet/requirements.txt
alembic==1.15.2
appdirs==1.4.4
apt-xapian-index==0.49
argcomplete==3.1.4
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.5
Brotli==1.1.0
build==1.0.3
cachetools==5.5.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.8
cloud-init==24.4.1
colorama==0.4.6
colour==0.1.5
command-not-found==0.3
configobj==5.0.8
coverage==7.4.4
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
discodos==1.1
DisplayCAL==0.0.0
distro==1.9.0
distro-info==1.7+build1
docopt==0.6.2
dropbox==11.36.2
easygui==0.98.1
Flask==3.1.0
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
fuse-python==1.0.7
git-cola==0.0.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
gpg==1.18.0
gphoto2==1.9.0
greenlet==3.1.1
gspread==6.2.0
gyp==0.1
helpdev==0.7.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
ifaddr==0.2.0
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libsass==0.22.0
louis==3.29.0
lxml==5.2.1
Mako==1.3.10
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.2.0
musicbrainzngs==0.7.1
mutagen==1.46.0
mysql-connector-python==9.2.0
mysqlx-connector-python==9.2.0
netaddr==0.8.0
netifaces==0.11.0
notify2==0.3
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
OpenLP==3.1.0rc4
packaging==24.0
pandas==2.2.3
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
polib==1.1.1
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyasyncore==1.0.2
pycairo==1.25.1
PyChromecast==9.4.0
pycryptodomex==3.20.0
pycups==2.0.1
pyenchant==3.2.2
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
pyinotify==0.9.6
PyJWT==2.7.0
pylibacl==0.7.0
pyliblo==0.10.0
pymediainfo==6.0.1
PyMuPDF==1.23.7
PyMySQL==1.1.1
pyodbc==5.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypng==0.20231004.0
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
pyrsistent==0.20.0
pyserial==3.5
pysword==0.2.8
pytest==7.4.4
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-vlc==3.0.20123
python3-discogs-client==2.5
pytz==2025.1
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==24.0.1
QDarkStyle==3.2.3
qrcode==7.4.2
QtAwesome==1.2.3
QtPy==2.4.1
qtsass==0.4.0
rapid-photo-downloader==0.9.36
rdflib==6.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9.1
scour==0.38.2
Send2Trash==1.8.2
setuptools==68.1.2
show-in-file-manager==1.1.4
simplejson==3.19.2
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.5
SPARQLWrapper==1.8.5
SQLAlchemy==2.0.40
stone==3.3.1
systemd-python==235
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
tornado==6.4
typing_extensions==4.13.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
wadllib==1.3.6
waitress==2.1.2
watchdog==3.0.0
webdavclient3==3.14.6
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.42.0
wxPython==4.2.1
xdg==5
xkit==0.0.0
yt-dlp==2024.4.9
zeroconf==0.131.0
